refactor extension reducer
