+++ packages/bierzo-wallet/src/store/reducers/extension/actions.ts
@@ -1,4 +1,4 @@
-import { SetExtensionStateActionType } from './types';
+import { SetExtensionStateActionType } from './reducer';
 
 export const setExtensionStateAction = (
   connected: boolean,
