+++ packages/bierzo-wallet/src/store/reducers/extension/index.ts
@@ -1,3 +1,2 @@
 export * from './actions';
 export * from './reducer';
-export * from './state';
