+++ packages/bierzo-wallet/src/store/reducers/extension/reducer.ts
@@ -1,6 +1,16 @@
 import { ActionType } from 'typesafe-actions';
 import * as actions from './actions';
-import { ExtensionState } from './state';
+import { Action } from 'redux';
+
+export interface ExtensionState {
+  readonly connected: boolean;
+  readonly installed: boolean;
+}
+
+export interface SetExtensionStateActionType extends Action {
+  type: '@@extension/SET_EXTENSION_STATE';
+  payload: ExtensionState;
+}
 
 export type ExtensionActions = ActionType<typeof actions>;
