--- packages/bierzo-wallet/src/store/reducers/extension/state.ts
@@ -1,4 +1,0 @@
-export interface ExtensionState {
-  readonly connected: boolean;
-  readonly installed: boolean;
-}
