--- packages/bierzo-wallet/src/store/reducers/extension/types.ts
@@ -1,7 +1,0 @@
-import { Action } from 'redux';
-import { ExtensionState } from './state';
-
-export interface SetExtensionStateActionType extends Action {
-  type: '@@extension/SET_EXTENSION_STATE';
-  payload: ExtensionState;
-}
